add-dish finished and created entity order

+++ src/main/java/com/waiter/waiter/entities/User.java
@@ -5,6 +5,7 @@ import com.waiter.waiter.enums.Role;
 import org.springframework.format.annotation.DateTimeFormat;
 
 import javax.persistence.*;
+import javax.persistence.criteria.CriteriaBuilder;
 import javax.validation.constraints.Email;
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Size;
@@ -17,7 +18,7 @@ public class User {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id", nullable = false, columnDefinition = "INT(11) UNSIGNED")
-    private Long id;
+    private Integer id;
 
     @NotNull
     @Size(min=5, max=30)
@@ -50,11 +51,11 @@ public class User {
     private Role role;//register input done
     private boolean enabled;
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
